+++ src/main/java/net/geforcemods/securitycraft/blocks/PanicButtonBlock.java
@@ -4,6 +4,8 @@ import net.geforcemods.securitycraft.api.OwnableBlockEntity;
 import net.geforcemods.securitycraft.misc.OwnershipEvent;
 import net.minecraft.block.BlockButton;
 import net.minecraft.block.ITileEntityProvider;
+import net.minecraft.block.material.EnumPushReaction;
+import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
@@ -45,6 +47,16 @@ public class PanicButtonBlock extends BlockButton implements ITileEntityProvider
 	@Override
 	public boolean isNormalCube(IBlockState state) {
 		return false;
+	}
+
+	@Override
+	public Material getMaterial(IBlockState state) {
+		return Material.ROCK;
+	}
+
+	@Override
+	public EnumPushReaction getPushReaction(IBlockState state) {
+		return EnumPushReaction.NORMAL;
 	}
 
 	@Override