+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedCobwebBlock.java
@@ -3,6 +3,7 @@ package net.geforcemods.securitycraft.blocks.reinforced;
 import javax.annotation.Nullable;
 
 import net.geforcemods.securitycraft.util.EntityUtils;
+import net.minecraft.block.SoundType;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
@@ -16,7 +17,8 @@ import net.minecraft.world.World;
 
 public class ReinforcedCobwebBlock extends BaseReinforcedBlock {
 	public ReinforcedCobwebBlock() {
-		super(Material.WEB, 1, Blocks.WEB);
+		super(Material.GLASS, 1, Blocks.WEB);
+		setSoundType(SoundType.STONE);
 	}
 
 	@Override